New Primary Vertex selection

+++ HiggsAnalysis/littleH/python/SimpleNtple_cff.py
@@ -46,7 +46,7 @@ def makeSimpleNtple(process, GlobalTag, MC=False, HLT='HLT', Filter=True, SavePA
     process.primaryVertexFilter = cms.EDFilter("GoodVertexFilter",
         vertexCollection = cms.InputTag('offlinePrimaryVertices'),
         minimumNDOF = cms.uint32(4) ,
-        maxAbsZ = cms.double(15),
+        maxAbsZ = cms.double(24),
         maxd0 = cms.double(2)
     )
